[REFACTORING] some decoupling of store-ui got lost
