+++ src/containers/auth-user-profile/redux-connect.js
