+++ src/containers/event-calendar/EventCalendar.js
@@ -4,7 +4,7 @@ import { Modal, Panel } from "react-bootstrap";
 import BigCalendar from "react-big-calendar";
 import "react-big-calendar/lib/css/react-big-calendar.css";
 
-import EventDetails from "../event-details/container";
+import EventDetails from "../event-details/redux-connect";
 
 let localizer = BigCalendar.momentLocalizer(moment);
 let allViews = Object.keys(BigCalendar.Views).map(k => BigCalendar.Views[k]);
