+++ src/containers/event-calendar/redux-connect.js
