+++ src/containers/event-guest-list/redux-connect.js
